Splits clik.magic into two separate modules.
